Fix playground links making editor un-editable (#134)
* Fix links making playground uneditable

The parsed program is now updated only once, when the page loads, rather
than when the component is updated.

* Revert "Fix links making playground uneditable"

This reverts commit 599db0f6c5e0673d704707aca92743439d369951.

* Standardize parameters for workspace

Now, we can pass in only a value and an optional library number.
This allows the Playground to handle parsing of the library program. If
there is a program passed in via the hash, the parsed program will be
passed. If not, the parsed program will be undefind, causing the redux
store value to be passed (using the `||` operator).

* Add test for playground with link

Also updated the old test snapshot

* Bump version 0.1.1 -> 0.1.2