Segregate Playground and Assessment Workspaces (#137)
* Remove updateEditorValue in workspace/DispatchProp

I will add this back as an OwnProps property, passed by either Assessment or
Playground

* Rename workspaceSaga to playgroundSaga

* Move actionTypes around

* Move playground actions into workspace/actions

Added an additional parameter to denote the location of the action call.

* Disconnect all workspace components

Due to the workspace's input and dispatch being determined by the
context, they can be passed in as props by Assessment or Playground.

* Refactor components that call Workspace

* Fill in missing props in Playground and Assessment

Also fixed some errors picked up at compilation time.

* Pass in SideContentProps instead of tabs

* Add location to saga calls

* Fix more errors

For some reason the member-ordering linting does not pass despite
proper(?) ordering. I'll open an issue for this.

* Modify default values for ControlBar

* Add WorkspaceManager state

This holds the 2 instances of IWorkspaceState, playground and
assessment. I've also added a scaffold of how the reducer function would
work, given a location. The location (type WorkspaceLocation) is now
specified to correspond to the key of the IWorkspaceState within the
IWorkspaceManagerState.

* Add location for reducer cases

Also made some interpreter actions to include location

* Fix PlaygroundContainer's IState reference

- Removed cached editorValue
- Renamed IState.workspace into workspaces
- Fixed a typo

* Add location to sagas

* Fix reducer naming error

* Rename workspace file to workspaces

* Add MapDispatchToProps call and DispatchProps

* Add dispatch and state references to container

* Use dispatch props in Playground component

Also added a missing handleChangeActiveTab

* Fix sideContentHeight not being stored in state

* Remove buggy currying function

For some reason, the spread and rest operators did not work with the
varying argument lengths. I proceeded to change them to a manual
currying. Also fixed a bug with the argument type for
handleEditorWidthChange.

* Fix property bug with evaluation

* Add state and dispatch for Assessment

* Sepatate Contexts for playground and assessment

* Format and Update test snapshots

* Use more elegant enums

* Sort properties

* Update test because of sorting

* Use editorValue in Assessment

* Make editorValue possibly undefined

This allows the "presetting" of values like in the case of a
solutionTemplate. However, it also forces any component to provide such
a value, thus I have added a template string for playground.

* Add .env to gitignore

* Add aciton to reset assessment workspace

* Add state slice for tracking assessment/question

* Format Assessment component

* Add comopnent props to get the current Assessment

* Add update method for current assessment

* Clean up logic for updating

* Fix logical errors preventing store changes

- For IStateProps, the properties I was assigning to were optional,
meaning that the linting did not catch the error. I overcame this by
storing and casting the return value.
- Made a mistake in using es6 object literal notation in the reducer, as
the property name was not as expected.

* Format and update tests

* Fix member ordering

* Fix sideContentHeight not auto-adjusting

* Rename saga

* Rename actions/workspace -> actions/workspaces

In line with the naming in the state and reducer.

* Fix playground default text bug

* Format some files